999 - Changed how benchmark ratings are selected.
In the old ratings, a location received 1 rating are we selected locations that were good or outstanding.
Now we have assessments, a location can have multiple ratings (per service), so how we select for good or outstanding needed to change.

I've added a numerical value for the overall_rating, then used a window to select locations where the min per location is >= 3 (good).