1162 - Changed impute_missing_values util function.
Added a loop through the schema then if the datatype of a column is a list, when that list is empty, replace it with null.

Before this change, the function would back/forward fill "non-null values" so empty lists were imputed.
In the pyspark pipeline, these were overwritten.
So i'm replicating the behaivour of the pyspark pipeline.